Slight refactors and additions to about us tests
