+++ webapp/src/components/other/DeveloperCards.tsx
@@ -1,9 +1,8 @@
-import * as React from 'react';
 import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
-import { Button, CardActionArea, CardActions, Grid } from '@mui/material';
+import { CardActionArea, Grid } from '@mui/material';
 import { Box } from '@mui/system';
 
 export default function MultiActionAreaCard() {
