+++ webapp/src/components/other/tests/DeveloperCards.test.tsx
@@ -1,10 +1,9 @@
-import React from 'react'
 import {render, screen } from '@testing-library/react';
-import DeveloperCards from './DeveloperCards';
+import DeveloperCards from '../DeveloperCards';
 
 test("DevelopersCard is rendered properly", async () =>{
 
-        render(<DeveloperCards></DeveloperCards>);
+        render(<DeveloperCards/>);
         expect(screen.getByText('Luis Miguel Alonso Ferreiro')).toBeInTheDocument();
         expect(screen.getByText('Sebastián López Hernández')).toBeInTheDocument();
         expect(screen.getByText('Jesús González Méndez')).toBeInTheDocument();
